GH-503 Separated states into files
