+++ SafeUIKit/ReplaceBrowserExtensionUI/RBEIntroViewControllerCancellingState.swift
@@ -1,0 +1,11 @@
+//
+//  Copyright © 2019 Gnosis Ltd. All rights reserved.
+//
+
+import Foundation
+
+extension RBEIntroViewController {
+
+    class CancellingState: State {}
+
+}
