+++ SafeUIKit/ReplaceBrowserExtensionUI/RBEIntroViewControllerStartedState.swift
@@ -1,0 +1,11 @@
+//
+//  Copyright © 2019 Gnosis Ltd. All rights reserved.
+//
+
+import Foundation
+
+extension RBEIntroViewController {
+
+    class StartedState: State {}
+
+}
